feat: 북마크 단위 테스트 수정
